Added my designs for sprint 2
These aren't very good, but they have some more functionality than the original. My main goal was to make it so it was more clear what spots you were placing your ships on by adding 1 through 10 on the side and A through J on the top of the battlefield. Sorry for the messy drawings, I am not very good at it :(